requirements.txt
pandas==2.2.3
mysql-connector-python==9.3.0
psycopg2-binary==2.9.10
dotenv~=0.9.9
python-dotenv~=1.1.0